Update to use app dispatch
