+++ lib/gateways/internal-api.ts
@@ -1,5 +1,4 @@
 import { AddressLookupResult } from '../../domain/addressLookup';
-import { NotifyRequest, NotifyResponse } from '../../domain/govukNotify';
 
 export const lookUpAddress = async (postCode: string) => {
   const res = await fetch(`/api/address/${postCode}`, {
@@ -7,13 +6,4 @@ export const lookUpAddress = async (postCode: string) => {
   });
 
   return (await res.json()) as AddressLookupResult;
-};
-
-export const sendNewApplicationEmail = async (notification: NotifyRequest) => {
-  const res = await fetch(`/api/notify/new-application`, {
-    method: 'POST',
-    body: JSON.stringify(notification),
-  });
-
-  return (await res.json()) as NotifyResponse;
 };
